fin-ops-pilot-agents/pyproject.toml
[project]
name = "fin-ops-pilot-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "fitz>=0.0.1.dev2",
    "livekit>=1.0.19",
    "openai>=2.8.1",
    "pydantic-settings>=2.12.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.38.0",
]
